--- src/Components/CiudadParticular/CiudadParticular.module.css
@@ -1,6 +1,0 @@
-.container {
-    display: flex;
-}
-.container li {
-    list-style-type: none;
-}
