--- src/Components/HotelParticular/HotelParticular.module.css
@@ -1,6 +1,0 @@
-.container {
-    display: flex;
-}
-.container li {
-    list-style-type: none;
-}
