+++ src/Routering.jsx
@@ -6,8 +6,8 @@ import LandPage from "./Pages/LandPage";
 import RegisterPage from "./Pages/RegisterPage";
 import CityViewPageAdmin from "./Pages/CityViewPageAdmin";
 import HotelViewPageAdmin from "./Pages/HotelViewPageAdmin";
-import CiudadParticular from "./Components/CiudadParticular/CiudadParticular";
-import HotelParticular from "./Components/HotelParticular/HotelParticular";
+import CiudadParticular from "./Components/Particular/CiudadParticular";
+import HotelParticular from "./Components/Particular/HotelParticular";
 import Feedback from "./Pages/FeedbackPage.jsx";
 import AddHab from "./Pages/AddHab.jsx";
 import UserProfile from "./Pages/UserProfile";
